+++ lib/resources_test.go
@@ -13,6 +13,7 @@ import (
 	"testing"
 	"time"
 
+
 	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation"
 	"github.com/aws/aws-sdk-go-v2/service/cloudformation/types"